+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 All notable changes to the "dependency-graph-viewer" extension will be documented in this file.
 
+### 0.0.3
+
+- Added javascript and javascript JSX support
+- Improved member functions matching in classes
+- Clicking on a File node opens the file
+- If the node is a reference of an exported entity, combine multiple occurrences to one node
+- Updated extension to open all files behind the scenes to make language server to find references from all files in the workspace
+added notification when loading
+
 ### 0.0.2
 
 Updated documentation.
 
 ### 0.0.1
 
-Initial release of TypeScript Dependency Graph Extension.
+Initial release of Dependency Graph Extension.